+++ src/main/java/net/nineocto/tossmod/TossMod.java
@@ -24,6 +24,8 @@ public class TossMod {
         modEventBus.addListener(this::commonSetup);
 
         MinecraftForge.EVENT_BUS.register(this);
+
+
     }
 
     private void commonSetup(final FMLCommonSetupEvent event) {
@@ -36,6 +38,7 @@ public class TossMod {
     public static class ClientModEvents {
         @SubscribeEvent
         public static void onClientSetup(FMLClientSetupEvent event) {
+
         }
     }
