+++ src/main/java/net/nineocto/tossmod/util/client/KeyBindings.java
@@ -1,16 +1,21 @@
 package net.nineocto.tossmod.util.client;
 
-import com.mojang.blaze3d.platform.InputConstants;
 import net.minecraft.client.KeyMapping;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
 import net.minecraftforge.client.event.RegisterKeyMappingsEvent;
-import net.minecraftforge.client.settings.KeyConflictContext;
 
+import static org.lwjgl.glfw.GLFW.GLFW_KEY_TAB;
+
+//This simply registers our keybind, It is triggered in Client Events.
 public class KeyBindings {
-    public static final String KEY_TOSS = "Drop: Toss Held Item";
-    public static KeyMapping tossKeyMapping;
-    public static void init(RegisterKeyMappingsEvent event) {
-        tossKeyMapping = new KeyMapping(KEY_TOSS, KeyConflictContext.IN_GAME,
-        InputConstants.getKey("key.keyboard.tab"), KeyMapping.CATEGORY_INVENTORY);
-        event.register(tossKeyMapping);
-    }
+    public static final String KEY_TOSS = "Drop Toss Held Item";
+   public static KeyMapping tossKeyMapping;
+        @OnlyIn(Dist.CLIENT)
+        public static void init(RegisterKeyMappingsEvent event) {
+            tossKeyMapping = new KeyMapping(
+                    KEY_TOSS, GLFW_KEY_TAB, "key.categories.inventory"
+            );
+            event.register(tossKeyMapping);
+        }
 }
