+++ src/main/java/net/nineocto/tossmod/util/network/tossmodPacketHandler.java
@@ -1,18 +1,15 @@
 package net.nineocto.tossmod.util.network;
 
-import net.minecraft.client.Minecraft;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraftforge.network.NetworkDirection;
-import net.minecraftforge.network.NetworkEvent;
 import net.minecraftforge.network.NetworkRegistry;
 import net.minecraftforge.network.PacketDistributor;
 import net.minecraftforge.network.simple.SimpleChannel;
 import net.nineocto.tossmod.TossMod;
 import net.nineocto.tossmod.util.network.packet.TossC2SPacket;
 
-import java.util.function.Supplier;
-
+//This is magic
 public class tossmodPacketHandler {
     private static SimpleChannel INSTANCE;
     private static int packetId = 0;
